refactor(etc/unittest): update async client test structure
